requirements.txt
django-cors-headers
Django>=5.2.3
djangorestframework